lsb-release: move to subiquitycore.os
