+++ subiquity/ui/views/help.py
@@ -21,7 +21,7 @@ from urwid import Divider, Filler, PopUpLauncher, Text, connect_signal
 from subiquity.common.types import PasswordKind
 from subiquity.ui.views.error import ErrorReportListStretchy
 from subiquitycore.async_helpers import run_bg_task
-from subiquitycore.lsb_release import lsb_release
+from subiquitycore.os import lsb_release
 from subiquitycore.ssh import summarize_host_keys
 from subiquitycore.ui.buttons import header_btn, other_btn
 from subiquitycore.ui.container import Columns, Pile, WidgetWrap
